Add support for whitelisting
makealerts checks against a whitelist lookup table with row indicating
criteria for whitelisting alerts.  If an alert is whitelisted, it is not
saved in the KVStore. Sufficient information should be available in the
super simple siem log to find out which rule was applied.

Rules are parsed using a parsec style combinator library, this should
prevent code injection (it does not use python's eval).

Added a Health dashboard with stats makealerts.