Switch from using mrsparkle framework to react.js
Up to Splunk 8.x, it was possible to mrsparkle application development
framework based on backbone.js and underscore.js. This has been
deprecated in 9.x. It requires to enable jQuery 3.5 and copy some
quarantine files which could present a security risk.

The new framework is based on react.js and is documented at
https://splunkui.splunk.com/home.

This commit allows Super Simple SIEM to work with Splunk 9.x without
enabling jQuery 3.5 and leverages the Splunk>ui react components.